refactor(coral): Add useEnvironmentTopics hook
